Vs analysis

+++ picoquic/intformat.c
@@ -107,7 +107,7 @@ void picoquic_varint_encode_16(uint8_t* bytes, uint16_t n16)
 
 size_t picoquic_varint_decode(const uint8_t* bytes, size_t max_bytes, uint64_t* n64)
 {
-    size_t length = 1 << ((bytes[0] & 0xC0) >> 6);
+    size_t length = ((size_t)1) << ((bytes[0] & 0xC0) >> 6);
 
     if (length > max_bytes) {
         length = 0;
@@ -128,7 +128,7 @@ size_t picoquic_varint_decode(const uint8_t* bytes, size_t max_bytes, uint64_t* 
 
 size_t picoquic_varint_skip(uint8_t* bytes)
 {
-    size_t length = 1 << ((bytes[0] & 0xC0) >> 6);
+    size_t length = ((size_t)1) << ((bytes[0] & 0xC0) >> 6);
 
     return length;
 }
